Bank Account Issue Resolved
There was an issue with the bank/betting, were when you placed a bet and won or lost you didn't get to bet more or bet less the next game it just reset itself to the default value of 100, this was a basic placement issue on my part where I placed it in a loop where once the game was over it when back to the second loop which called the bank class that reset the total value.e Now by moving it outside of the loops it's corrected and you can bet as much as you want as long as you can win. I plan on creating a message that will pop up when you run out of money, but it hasn't been implemented yet but will be soon.