Fixed bet amount adding issue
Fixed an issue regarding the bet amount constantly adding on top of each other, this was due to not resetting the self.bet amount to 0 every time we called the make_bet method. Since before the make_bet method didn't self.bet = 0 it was just adding on top of the previous bet amount. To fix this, I just had to put self.bet = 0 so that it doesn't keep adding on top of the self.bets the previous amount. For example, if previously you bet 100 and won, your total would be 200, and if next round you wanted to bet 50, it would add 50 to 150 equallying 200, and add that 200 to the total which would be 200, so instead of the total being 250 it would be 400.